backend/requirements.txt
flask==1.1.2
Flask-Cors==3.0.9
python-dotenv==0.14.0
requests==2.11.1
Pyrebase==3.0.27
Cerberus==1.3.2